% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningHelpers.R
\name{checkVars}
\alias{checkVars}
\title{checkVars}
\usage{
checkVars(data, varNames, except = NULL)
}
\arguments{
\item{data}{Raw data, a data frame.}

\item{varNames}{Vector of valid variable names.}

\item{except}{A vector of variable names to be excluded form the check.}
}
\value{
A list.
}
\description{
Check raw data: are all required variables preseent and 
    ensure there are no extraneous variables.
}
