% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningHelpers.R
\name{checkVarsExist}
\alias{checkVarsExist}
\title{checkVarsExist}
\usage{
checkVarsExist(data, varNames)
}
\arguments{
\item{data}{Raw data, a data frame.}

\item{varNames}{Vector of valid variable names.}
}
\value{
A list.
}
\description{
Check that all required variables exist in raw data.
}
\keyword{internal}
