% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestr.R
\docType{package}
\name{conquestr}
\alias{conquestr}
\alias{conquestr-package}
\title{conquestr: A package for exchanging information with ConQuest software}
\description{
conquestr: A package for exchanging information with ConQuest software
}
\section{The foo package provides two categories of important functions}{

 Functions for exporting information to ConQuest, and
 Functions for importing ConQuest output
}

\section{export functions}{
 The main function is `cq_analysis` which is used to prepare ConQuest command syntax and files for
  analysis based on a supplied dataframe.
}

\section{import functions}{
 These functions import the results of a ConQuest analysis.  These include
  `cq_itanal` to import item analysis files
  `cq_pv` to import plausible values files
}

\references{
Wu, M., Adams, R., Wilson, M., & Haldane, S. (2007). ConQuest Version 2: Generalised Item Response Modelling Software.
}
