% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestr.R
\name{ConQuestRout}
\alias{ConQuestRout}
\title{ConQuestRout}
\usage{
ConQuestRout(myRout)
}
\arguments{
\item{myRout}{The path to the binary rout file.}
}
\value{
A list containing the data objects to recreate 'ACER ConQuest' plots.
}
\description{
Read an 'ACER ConQuest' rout file (a binary file)
created by a \href{https://conquestmanual.acer.org/s4-00.html#commandPlot}{\code{plot}}
command in 'ACER ConQuest'.
}
\examples{
myPlot <- ConQuestRout()
\dontrun{
# the default example is an ICC plot from Example 1 
# <https://conquestmanual.acer.org/s2-00.html#s2-02>.
str(myPlot)
}
}
