% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_example.R
\name{cq_example}
\alias{cq_example}
\title{Example ConQuest output files}
\usage{
cq_example(n = 10, display = TRUE, example_name = c("ex1.itn",
  "ex1_05.pv", "ex1_10.pv"))
}
\arguments{
\item{n}{the number of lines to print. only used when \code{display = TRUE}.}

\item{display}{When \code{TRUE} (the default) the example output is read and the first \code{n} lines are printed.
The file path to example conquest output file is returned invisibly.
Use \code{FALSE} if just the conquest output file path is needed. No output will be displayed.}

\item{example_name}{Specifies the example output file to use. Default is \code{"ex1.itn"}. Other options are
\code{"ex1_05.pv"} and \code{"ex1_10.pv"} which requests the plausible values example file with 5 or 10 pvs per case respectively.}
}
\value{
Invisibly returns a valid file path to conquest output file which can be used with other conquestr functions.
}
\description{
Example ConQuest output files
}
\examples{
library(readr)
fname <- cq_example(display = FALSE)
fname
cq_example(n = 7)

}
