% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_example.R
\name{cq_example_itanal}
\alias{cq_example_itanal}
\title{cq_example_itanal (deprecated)}
\usage{
cq_example_itanal(...)
}
\arguments{
\item{...}{arguments passed to \code{cq_example}}
}
\description{
\code{cq_example_itanal} has been deprecated and may be removed from future versions.
Please use \code{cq_example} instead.
}
