% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_itanal.R
\name{cq_itanal}
\alias{cq_itanal}
\title{Main function to read an itanal. If it doesn't work try using cq_read_itanal instead}
\usage{
cq_itanal(fname)
}
\arguments{
\item{fname}{A file path to an existing ConQuest itanal file. Can take other input as for readr::read_file}
}
\value{
A tibble containing itanal contents with numeric fields where appropriate.
}
\description{
Main function to read an itanal. If it doesn't work try using cq_read_itanal instead
}
\examples{
fname <- cq_example(display = FALSE)
df <- cq_itanal(fname)

}
