% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_pv.R
\name{cq_pv}
\alias{cq_pv}
\title{Read in a ConQuest produced plausible values file}
\usage{
cq_pv(fname, np = 5)
}
\arguments{
\item{fname}{A file path to an existing ConQuest plausible values file. Can take other input as for readr::read_file}

\item{np}{Number of plausible values per case. Default is 5.}
}
\value{
A tibble containing parsed results of the file.
}
\description{
Read in a ConQuest produced plausible values file
}
\examples{
fname <- cq_example(display = FALSE, example_name = "ex1_05.pv")
df <- cq_pv(fname)

}
\seealso{
\code{\link[sirt]{read.pv}} which also reads plausible values files. For reading in multidimensional pv files try
\code{\link[sirt]{read.multidimpv}} or \url{https://rdrr.io/rforge/eatRest/src/R/get.plausible.R}
}
