% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_itanal.R
\name{cq_read_itanal}
\alias{cq_read_itanal}
\title{Read in a traditional ConQuest item analysis file as text}
\usage{
cq_read_itanal(fname)
}
\arguments{
\item{fname}{A file path to an existing ConQuest itanal file. Can take other input as for readr::read_file}
}
\value{
A tibble containing parsed results of the itanal file.
}
\description{
Read in a traditional ConQuest item analysis file as text
}
\examples{
fname <- cq_example_itanal(display = FALSE)
df <- cq_read_itanal(fname)

}
