% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_itanal.R
\name{cq_resp_stats}
\alias{cq_resp_stats}
\title{Extracts the response stats from an itanal into usable form}
\usage{
cq_resp_stats(itanal)
}
\arguments{
\item{itanal}{An itanal tibble returned by cq_read_itanal or cq_itanal}
}
\value{
A tibble containing stats for each item response
}
\description{
Extracts the response stats from an itanal into usable form
}
\examples{
fname <- cq_example_itanal(display = FALSE)
df <- cq_itanal(fname)
cq_resp_stats(df)

}
