% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_show.R
\name{cq_show}
\alias{cq_show}
\title{Read in ConQuest show output file}
\usage{
cq_show(fname)
}
\arguments{
\item{fname}{A file path to an existing ConQuest show file (e.g.
'myfile.shw'). Can take other input as for \code{readr::read_file}}
}
\description{
Read in ConQuest show output file
}
\examples{
fname <- cq_example(display = FALSE, example_name = "ex1.shw")
df <- cq_show(fname)

}
