% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_syntax.R
\name{cqc_cmds}
\alias{cqc_cmds}
\title{List of ConQuest commands}
\usage{
cqc_cmds()
}
\value{
A named list. Each element contains a string containing a ConQuest syntax statement with placeholders
}
\description{
List of ConQuest commands
}
\examples{
cqc_cmds()

}
