% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_syntax.R
\name{cqc_defaults}
\alias{cqc_defaults}
\title{List of default arguments for use with ConQuest commands}
\usage{
cqc_defaults()
}
\value{
A named list. Each element contains a default placeholder value. Designed for use with `cqc_cmds`
}
\description{
List of default arguments for use with ConQuest commands
}
\examples{
cqc_defaults()

}
