% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_syntax.R
\name{cqc_exact_wrap}
\alias{cqc_exact_wrap}
\title{exactly wrap text at the width specified}
\usage{
cqc_exact_wrap(some_text, width = 80)
}
\arguments{
\item{some_text}{A string. (i.e. character vector of length 1)}

\item{width}{The character width to wrap at}
}
\value{
A string wrapped at exactly \code{width} intervals
}
\description{
\code{strwrap()} in base wraps at word boundaries. \code{cqc_exact_wrap()} is
for when wrapping at the exact width is needed.
}
\examples{
x <- cqc_exact_wrap("abcde", 2)
x
cat(x)

}
