% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_syntax.R
\name{cqc_resp_cols}
\alias{cqc_resp_cols}
\title{Condense ConQuest response columns syntax}
\usage{
cqc_resp_cols(x)
}
\arguments{
\item{x}{takes an integer vector (or attempts to coerce to one) which
represents column positions for ConQuest items in a fixed-width text file
used for analysis.}
}
\value{
a string containing response columns for use with the ConQuest format statement
}
\description{
Designed to work with \code{cqc_data}. This is a convenience functions
designed for use when a subset of items are being analysed that are contained
in the ConQuest data file.
}
\examples{
x <- c(1,3,4,5,7,9,10,11)
cqc_resp_cols(x)

}
