% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_syntax.R
\name{cqc_syntax}
\alias{cqc_syntax}
\title{Generate ConQuest syntax, with some reasonable defaults}
\usage{
cqc_syntax(name, resp_cols, cmds = cqc_cmds(), lookup_vals = cqc_defaults())
}
\arguments{
\item{name}{A name for the analysis - this is used to link data, labels and output files}

\item{resp_cols}{Columns containing item responses. For use with ConQuest format statment}

\item{cmds}{A list of strings. Each string containing a ConQuest statment (and may include placeholders)}

\item{lookup_vals}{A named list of strings. The name defines a 'placeholder' for use with ConQuest commands. The string is the value that will be used instead of the placeholder}
}
\value{
A named list. Each element contains a Conquest command. Use with `cqc_cmds` and `cqc_defaults`
}
\description{
Generate ConQuest syntax, with some reasonable defaults
}
\examples{
# cqc_syntax only generates text which can be used by ConQuest. e.g.
cqc_syntax(name = "mydata", resp_cols = "1-10")


}
