% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{createConQuestProject}
\alias{createConQuestProject}
\title{createConQuestProject}
\usage{
createConQuestProject(prefix = getwd(), ...)
}
\arguments{
\item{prefix}{a valid file path where to create project folders.}

\item{...}{optional params, including "setDebug"}
}
\value{
Boolean TRUE.
}
\description{
creates a standard folder structure to work with 'ACER ConQuest' Projects.
}
\examples{
\dontrun{
createConQuestProject()
}
}
