% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState_createDF.R
\name{createDfFromSys}
\alias{createDfFromSys}
\title{createDfFromSys}
\usage{
createDfFromSys(mySys)
}
\arguments{
\item{mySys}{An R object of class ConQuestSys, returned by the function conquestr::ConQuestSys}
}
\value{
A data frame containing R data frames based on the list objects in the ConQuest system file that has been read in.
}
\description{
Read an R object of class ConQuestSys and create neat R data frame object.
}
\seealso{
conquestr::ConQuestSys()
}
