% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState.R
\name{DecompressSys}
\alias{DecompressSys}
\title{DecompressSys}
\usage{
DecompressSys(myFile)
}
\arguments{
\item{myFile}{An connection to a compressed  'ACER ConQuest' system file
created by the \code{put} command in 'ACER ConQuest'.`}
}
\value{
A connection to an uncompressed system file in the users temp dir.
}
\description{
Internal function to decompress an 'ACER ConQuest' system file
that has been compressed using zlib.
}
\seealso{
conquestr::ConQuestSys()
}
