% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{findConQuestExe}
\alias{findConQuestExe}
\title{findConQuestExe}
\usage{
findConQuestExe()
}
\value{
Char with path to ConQuest executable.
}
\description{
Searches in common insall paths to find ConQuest executable.
This is called by `ConQuestCall` when no executable is passed explicitly.
}
\examples{
\dontrun{
findConQuestExe()
}
}
\keyword{internal}
