% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residHelpers.R
\name{fisherTrnsfrm}
\alias{fisherTrnsfrm}
\title{fisherTrnsfrm}
\usage{
fisherTrnsfrm(myCorMat)
}
\arguments{
\item{myCorMat}{A correlation matrix.}
}
\value{
A correlation matrix with Fisher's transform applied to values -1 > x > 1.
}
\description{
Helper function to apply Fisher's transformation to a correlation matrix.
}
