% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itanalHelpers.R
\name{fmtCqItanalTbl}
\alias{fmtCqItanalTbl}
\title{fmtCqItanalTbl}
\usage{
fmtCqItanalTbl(
  cqItanalTbl,
  ptBisFlag,
  textColHighlight,
  valueDecPlace,
  itemName
)
}
\arguments{
\item{cqItanalTbl}{A single table in a cqItanal object (an element of type data.frame in a list of class "cqItanal".}
}
\value{
A data frame.
}
\description{
internal function to produce formatted itanal category statistics tables for a generalised item.
Note that this function is called by \code{fmtCqItanal} and would not usually be called by the user.
The returned table uses the library kable and will be formatted
based on the users settings.
}
\keyword{internal}
