% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateItems.R
\name{genItems_deltaDots}
\alias{genItems_deltaDots}
\title{genItems_deltaDots}
\usage{
genItems_deltaDots(deltadots, n)
}
\arguments{
\item{deltadots}{A list that describes the sampling distribution from which
item location parameters are drawn.}

\item{n}{An integer describing the number of items being generated.}
}
\value{
A vector of delta dots.
}
\description{
Internal function to parse deltaDots argument
to \code{conquestr::genItems}.
}
\keyword{internal}
