% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateItems.R
\name{genItems_taus}
\alias{genItems_taus}
\title{genItems_taus}
\usage{
genItems_taus(taus, n)
}
\arguments{
\item{taus}{A list that describes the sampling distribution from which
item location parameters are drawn.}

\item{n}{An integer describing the number of items being generated.}
}
\value{
A list
}
\description{
Internal function to parse taus argument
to \code{conquestr::genItems}.
}
\keyword{internal}
