% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{geom_line_or_smooth}
\alias{geom_line_or_smooth}
\title{geom_line_or_smooth}
\usage{
geom_line_or_smooth(thisData, thisLinetype)
}
\arguments{
\item{thisData}{the data frame being worked on in the call to \code{conquestr::plotCCC()}}

\item{thisLinetype}{a string, either "bins" or "regression" set in the call to \code{conquestr::plotCCC()}}
}
\value{
a list with a function call to \code{ggplot2::geom_line}.
}
\description{
helper functions for ggplot - conditionally creates call to
\code{ggplot2::geom_smooth} depending on the option linetype in the call to \code{conquestr::plotCCC()}
}
\examples{
myExample <- geom_line_or_smooth()
}
\keyword{internal}
