% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState_createDF.R
\name{getCqEsts}
\alias{getCqEsts}
\title{getCqEsts}
\usage{
getCqEsts(mySys)
}
\arguments{
\item{mySys}{An R object of class conQuestSysFile, returned by the function conquestr::ConQuestSys}
}
\value{
A data frame containing ability estimates (missing if not estimated).
}
\description{
Return ability estimates as a data frame.
}
\keyword{internal}
