% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqFit}
\alias{getCqFit}
\title{getCqFit}
\usage{
getCqFit(myCqs)
}
\arguments{
\item{myCqs}{A system file.}
}
\value{
A data frame.
}
\description{
creates a data frame representation of the fit of parameters in the item reponse model
}
\examples{
\dontrun{
getCqFit(ConQuestSys())
}
}
