% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqHist}
\alias{getCqHist}
\title{getCqHist}
\usage{
getCqHist(myCqs)
}
\arguments{
\item{myCqs}{A system file.}
}
\value{
A data frame.
}
\description{
creates a data frame representation of the iteration history for all parameters.
}
\examples{
\dontrun{
getCqHist(ConQuestSys())
}
}
