% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itanalHelpers.R
\name{getCqItanal}
\alias{getCqItanal}
\title{getCqItanal}
\usage{
getCqItanal(sysFile, matrixPrefix = "", isDebug = FALSE)
}
\arguments{
\item{sysFile}{An ACER ConQuest system file.}

\item{matrixPrefix}{to define which itanal to process}

\item{isDebug}{report debug output}
}
\value{
A list.
}
\description{
helper function to return list of lists, each list relates to one 
  generalised item from an ACER ConQuest `itanal` output.
  Each list contains: (1) item-total and item-rest correlations ....
}
\examples{
myItanal <- getCqItanal()
print(myItanal[[1]])
}
