% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itanalHelpers.R
\name{getCqItanalSummary}
\alias{getCqItanalSummary}
\title{getCqItanalSummary}
\usage{
getCqItanalSummary(itan)
}
\arguments{
\item{itan}{A list of class "cqItanal" created by \code{conquestr::getCqItanal()}}
}
\value{
A data frame.
}
\description{
returns an itanal as a data frame in summary format:
one row per generalised item with:
\itemize{
\item item label
\item valid N
\item facility (see \code{conquestr::getCqItanalFacility})
\item item-rest correlation
\item item-total correlation
\item fit (infit/weighted MNSQ) if available
\item item locations (deltas)
}
}
\examples{
mySys <- ConQuestSys()
myItan <- getCqItanal(mySys)
getCqItanalSummary(myItan)
}
