% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqLongLabs}
\alias{getCqLongLabs}
\title{getCqLongLabs}
\usage{
getCqLongLabs(sysFile)
}
\arguments{
\item{sysFile}{An ACER ConQuest system file read into R using conquestr::ConQuestSys.}
}
\value{
A vector.
}
\description{
returns a vector of long gin labels returns vector of length 0
if there are no labels used (see command labels in conquest)
}
\examples{
\dontrun{
getCqLongLabs(ConQuestSys())
}
}
\keyword{internal}
