% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqParams}
\alias{getCqParams}
\title{getCqParams}
\usage{
getCqParams(myCqs)
}
\arguments{
\item{myCqs}{A system file.}
}
\value{
A data frame.
}
\description{
creates a data frame representation of the parameters of the model, including both estimated and constrained parameters.
   Parameters are either freely estimated ('ParamType' == 0) or constrained ('ParamType' == 0).
   Parameters are indexed (0 offset) by the column 'ParamNumber'. There is a seperate index for free and constrained parameters.
}
\examples{
\dontrun{
getCqParams(ConQuestSys())
}
}
