% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqParams2}
\alias{getCqParams2}
\title{getCqParams2}
\usage{
getCqParams2(myCqs)
}
\arguments{
\item{myCqs}{A system file.}
}
\value{
A list of data frames of length `gTerms`.
}
\description{
returns a list of data frames, one for each term in the model.
   The data frames include the the estimate, location constraint, or anchored value for each parameter.
   The standard error of estimated parameters is also included if it has been estimated.
   The fit (un/weighted MNSQ, sometimes called outfit and infit) of all parameters is alre included if has been estimated.
}
\examples{
\dontrun{
getCqParams2(ConQuestSys())
}
}
