% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState_createDF.R
\name{getCqPid}
\alias{getCqPid}
\title{getCqPid}
\usage{
getCqPid(mySys)
}
\arguments{
\item{mySys}{An R object of class conQuestSysFile, returned by the function conquestr::ConQuestSys}
}
\value{
A data frame containing sequence number and PID (if no PID is declared, this is the sequence number).
}
\description{
Return PID as a data frame.
}
\keyword{internal}
