% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState_createDF.R
\name{getCqResp}
\alias{getCqResp}
\title{getCqResp}
\usage{
getCqResp(mySys)
}
\arguments{
\item{mySys}{An R object of class conQuestSysFile, returned by the function conquestr::ConQuestSys}
}
\value{
A data frame containing raw item responses (pre key) and scored item response (post key).
}
\description{
Return item responses as a data frame.
}
\keyword{internal}
