% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itanalHelpers.R
\name{getCqScores}
\alias{getCqScores}
\title{getCqScores}
\usage{
getCqScores(sysFile)
}
\arguments{
\item{sysFile}{An ACER ConQuest system file.}
}
\value{
A data frame.
}
\description{
internal function to return scores associated with item categories in itanal tables.
}
\keyword{internal}
