% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqTerms}
\alias{getCqTerms}
\title{getCqTerms}
\usage{
getCqTerms(myCqs)
}
\arguments{
\item{myCqs}{A system file.}
}
\value{
A data frame.
}
\description{
creates a data frame representation of the terms of the model statement, including interactions.
}
\examples{
\dontrun{
getCqTerms(ConQuestSys())
}
}
