% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState_createDF.R
\name{getCqYData}
\alias{getCqYData}
\title{getCqYData}
\usage{
getCqYData(mySys)
}
\arguments{
\item{mySys}{An R object of class conQuestSysFile, returned by the function conquestr::ConQuestSys}
}
\value{
A data frame containing weight and regression data.
}
\description{
Return weight and regression data as a data frame.
}
\keyword{internal}
