% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showHelpers.R
\name{ginsOnDims}
\alias{ginsOnDims}
\title{ginsOnDims}
\usage{
ginsOnDims(sysFile)
}
\arguments{
\item{sysFile}{An ACER ConQuest system file read into R using conquestr::ConQuestSys}
}
\value{
a list
}
\description{
returns a list of length gNDims. Each element of the list contains
a vector of the gins on this dim.
}
\examples{
\dontrun{
myResult <- ginsOnDims(conquestr::ConQuestSys())
}
}
