% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{guide_remove_linetype}
\alias{guide_remove_linetype}
\title{guide_remove_linetype}
\usage{
guide_remove_linetype(groups = TRUE)
}
\arguments{
\item{groups}{A bool indicating if there is groups in the call to \code{conquestr::plotCCC()}.}
}
\value{
a list with a call to \code{ggplot2::guides}.
}
\description{
helper functions for ggplot - conditionally creates call to
\code{ggplot2::guides} to remove legends that involve groups when there is
no groups in the call to \code{conquestr::plotCCC()}
}
\examples{
myExample <- guide_remove_linetype(TRUE)
}
\keyword{internal}
