% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{isCqConverged}
\alias{isCqConverged}
\title{isCqConverged}
\usage{
isCqConverged(myCqs)
}
\arguments{
\item{myCqs}{A system file.}
}
\value{
A boolean.
}
\description{
returns true is the ConQuest model has converged normally (system file).
}
\examples{
\dontrun{
isCqConverged(ConQuestSys())
}
}
\keyword{internal}
