% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{ItemDfStdZ}
\alias{ItemDfStdZ}
\title{ItemDfStdZ}
\usage{
ItemDfStdZ(myDf)
}
\arguments{
\item{myDf}{a data frame.}
}
\value{
a data frame.
}
\description{
Calculates centred and scaled item parameter estimates. Also calculates standardised standard errors of item parameter estimates to complement scaled item parameter estimates.
}
\seealso{
conquestr::sysToItemDifDf()
}
\keyword{internal}
