% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{ItemDfStdZMerge}
\alias{ItemDfStdZMerge}
\title{ItemDfStdZMerge}
\usage{
ItemDfStdZMerge(myDf1, myDf2, mySuffixes)
}
\arguments{
\item{myDf1}{a data frame.}

\item{myDf2}{a data frame.}

\item{mySuffixes}{a vector of strings specifying the names for the two groups being analysed, e.g., if the two system files are an analysis of boys and girls, the vector may be `c(_male", "_female")`.}
}
\value{
a data frame.
}
\description{
Calculates Z test on the common items from 2 data frames returned from `conquestr::ItemDfStdZ`.
}
\seealso{
conquestr::sysToItemDifDf()
}
\keyword{internal}
