% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoHelpers.R
\name{itemInfoAtTheta}
\alias{itemInfoAtTheta}
\title{itemInfoAtTheta}
\usage{
itemInfoAtTheta(myItem, theta)
}
\arguments{
\item{myItem}{A matrix of item parameters of the structure used in `simplef`}

\item{theta}{a number.}
}
\description{
Calculates item information at a value of theta given a set of item parameters.
}
\examples{
anItem<-  matrix(c(0,0,0,1,1,0), nrow = 2, byrow = TRUE)
itemInfoAtTheta(anItem, 0)
}
