% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrstThrsh.R
\name{itemListToDeltaDots}
\alias{itemListToDeltaDots}
\title{itemListToDeltaDots}
\usage{
itemListToDeltaDots(myItems)
}
\value{
A data frame including 4 columns:
\itemize{
\item id, an integer index reflecting which item this is, in the same order as myItems
\item itemid, a string with the names from the items in myItems (NA if item list is not named)
\item location, the value of the deta dot
}
}
\description{
Taskes a list of item parameter matricies and returns a data frame containing
delta dots. This is mostly an internal function used, for example to plot items
on Wright Maps.
}
\examples{
myItem <- matrix(
  c(
    0, -0.58    ,  0     , 1,  # delta+tau   thurst thresh (gamma)
    1, -0.58    ,  0.776 , 1,  # 0.196       -1.14
    2, -0.58    , -0.697 , 1,  # -1.277      -0.93
    3, -0.58    , -0.629 , 1,  # -1.209      -0.64
    4, -0.58    ,  0.55  , 1   # -0.03        0.25
    ), ncol =4, byrow=TRUE
 )
thrstThrsh(myItem)
}
\keyword{internal}
