% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoHelpers.R
\name{myAbilitiesHelper}
\alias{myAbilitiesHelper}
\title{myAbilitiesHelper}
\usage{
myAbilitiesHelper(myWhat = "ok", myAbilities = NA)
}
\arguments{
\item{myWhat}{A character string indicating what to check.}

\item{myAbilities}{A numeric vector of person abilities.}
}
\value{
A Boolean or a double.
}
\description{
helper function to return attributes of vector of abilities for use in information function and WI.
}
\keyword{internal}
