% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoHelpers.R
\name{myItemHelper}
\alias{myItemHelper}
\title{myItemHelper}
\usage{
myItemHelper(myWhat = "ok", myItems)
}
\arguments{
\item{myWhat}{A character string indicating what to check.}

\item{myItems}{A list of item matricies.}
}
\value{
A Boolean or a double.
}
\description{
helper function to return attributes of list of items for use in information function helpers and WI.
}
\keyword{internal}
