% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itanalHelpers.R
\name{orderItanalCols}
\alias{orderItanalCols}
\title{orderItanalCols}
\usage{
orderItanalCols(itanalTbl)
}
\arguments{
\item{itanalTbl}{An itanal table being worked on by fmtCqItanalTbl.}
}
\value{
A data frame.
}
\description{
internal function to order and rename itanal tables.
Called by conquestr::fmtCqItanalTbl
Safely orders and renames tables when ndim > 1.
}
\keyword{internal}
