% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{plotModelExp}
\alias{plotModelExp}
\title{plotModelExp}
\usage{
plotModelExp(items, range = c(-6, 6), by = 0.1)
}
\arguments{
\item{items}{List of one or more matricies of item parameters.}

\item{range}{Lower and upper bounds to plot over (defaults to c(-6, 6).}

\item{by}{Increment to the sequence along `range``.}
}
\value{
A ggplot2 object.
}
\description{
Creates a plot of a model-implied expected score curve.
Note this is not for use with \code{rout} files. See the generic function \code{plotRout} for plotting rout files.
}
\examples{
myItem <- matrix(
  c(
    0, 0, 0, 1,
    1, 1, 0, 1
  ),
  ncol = 4, byrow=TRUE
)
myPlot <- plotModelExp(list(myItem))
}
