% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessCqSysfile.R
\name{postProcess_gLConstraint}
\alias{postProcess_gLConstraint}
\title{postProcess_gLConstraint}
\usage{
postProcess_gLConstraint(this_gLConstraint, debug)
}
\arguments{
\item{this_gLConstraint}{An object of type gLConstraint from a list of class conQuestSysFile}

\item{debug}{is this a debug run?}
}
\value{
A list containing a documented gLConstraint object
}
\description{
internal function to process gLConstraint
}
\seealso{
conquestr::postProcessSys()
}
\keyword{internal}
