% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessCqSysfile.R
\name{postProcessGncases}
\alias{postProcessGncases}
\title{postProcessGncases}
\usage{
postProcessGncases(this_gNCases, debug)
}
\arguments{
\item{this_gNCases}{An object of type gNCases from a list of class conQuestSysFile}

\item{debug}{is this a debug run?}
}
\value{
A list containing a documented gNCases object
}
\description{
internal function to processing gNCases
}
\seealso{
conquestr::postProcessSys()
}
\keyword{internal}
