% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessCqSysfile.R
\name{postProcessSys}
\alias{postProcessSys}
\title{postProcessSys}
\usage{
postProcessSys(sysFile)
}
\arguments{
\item{myFile}{A list of class conQuestSysFile}
}
\value{
A list containing the data objects created by 'ACER ConQuest'.
Object is of class conQuestSysFile.
}
\description{
Post processing of ConQuest system file created by ConQuestSys.
Adds names and attributes (e.g., descriptions) to objects in list.
}
\seealso{
conquestr::ConQuestSys()
}
\keyword{internal}
