% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvHelpers.R
\name{pvMeanVar}
\alias{pvMeanVar}
\title{pvMeanVar}
\usage{
pvMeanVar(myData)
}
\arguments{
\item{myData}{A matrix of PVs for one dimension: m PVs by n cases.}
}
\value{
A list containing the mean and variaince of the PVs.
}
\description{
Applies the law of total variance (EVEs law)
    to calculate the mean and varinace of a set of PVs for one dimension.
}
