% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residHelpers.R
\name{q3ExpCorrect}
\alias{q3ExpCorrect}
\title{q3ExpCorrect}
\usage{
q3ExpCorrect(myCorMat)
}
\arguments{
\item{myCorMat}{A correlation matrix.}
}
\value{
A correlation matrix with the Q3 statistic correction applied.
}
\description{
Helper function to apply correction to correlation matrix. When working with standardised residuals,
the expectation of the correlations is -1/(L-1) rather than 0 See DOI: 10.1177/0013164410379322
}
