% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHelpers.R
\name{rawScore}
\alias{rawScore}
\title{rawScore}
\usage{
rawScore(x, itemParams, perfAdjust = 0.3)
}
\arguments{
\item{x}{a vector of scored item responses. Order of item responses is the same as the order of \code{itemParams},
missing item responses are scored NA.}

\item{itemParams}{A list of item params of the structure used in \code{simplef} (a matrix of k categories by three
(category score, delta dot, tau)). See conquestr::makeItemList for a helper to generate this list.}

\item{perfAdjust}{the correction factor for zeros and perfects to be add/subtracted from the raw score.}
}
\value{
a list containing three elements: (1) raw score (unadjusted), (2) raw score (adjusted), (3) maximum score.
}
\description{
returns the adjusted (for zero and perfects) and unadjusted raw scores as well as the maximum score.
}
\keyword{internal}
