% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadADesignMatrices}
\alias{ReadADesignMatrices}
\title{ReadADesignMatrices}
\usage{
ReadADesignMatrices(myFile, Columns, Items, ItemSteps)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}

\item{Columns}{An integer representation of 'ACER ConQuest' object gNParameters.}

\item{Items}{An integer representation of 'ACER ConQuest' object gNGins.}

\item{ItemSteps}{An integer representation of 'ACER ConQuest' object gItemSteps.}
}
\value{
A list
}
\description{
ReadSys Read the A design matrix (A list of length gNGins of matrices.  For each matrix the number of rows is gItemSteps (for that item) and the number of columns is gNParameters_C).
}
\keyword{internal}
