% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadAllCaseEstimates}
\alias{ReadAllCaseEstimates}
\title{ReadAllCaseEstimates}
\usage{
ReadAllCaseEstimates(myFile, Dimensions, N, NPlausibles)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}

\item{Dimensions}{An integer representation of 'ACER ConQuest' object gNDim.}

\item{N}{An integer representation of 'ACER ConQuest' object gNCases}

\item{NPlausibles}{An integer representation of 'ACER ConQuest' object gNPlausibles.}
}
\value{
A list
}
\description{
ReadAllCaseEstimates
}
\keyword{internal}
