% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadAllGroupsData}
\alias{ReadAllGroupsData}
\title{ReadAllGroupsData}
\usage{
ReadAllGroupsData(myFile, N, GroupVariables, AllVariables)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}

\item{N}{An integer representation of 'ACER ConQuest' object gNCases.}

\item{GroupVariables}{A list of group variables.}

\item{AllVariables}{A list of variables.}
}
\value{
A list
}
\description{
ReadAllGroupsData
}
\keyword{internal}
