% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadAllResponseData}
\alias{ReadAllResponseData}
\title{ReadAllResponseData}
\usage{
ReadAllResponseData(myFile, N)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}

\item{N}{An integer representation of 'ACER ConQuest' object gNCases.}
}
\value{
A list
}
\description{
ReadAllResponseData
}
\keyword{internal}
