% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadAllY}
\alias{ReadAllY}
\title{ReadAllY}
\usage{
ReadAllY(myFile, N, NReg)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}

\item{N}{An integer representation of 'ACER ConQuest' object gNCases}

\item{NReg}{An integer representing the number of regressors in the model, a representation of 'ACER ConQuest' object gNReg.}
}
\description{
ReadAllY
}
\keyword{internal}
