% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadBandDefinesList}
\alias{ReadBandDefinesList}
\title{ReadBandDefinesList}
\usage{
ReadBandDefinesList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadBandDefinesList
}
\keyword{internal}
