% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadBoolean}
\alias{ReadBoolean}
\title{ReadBoolean}
\usage{
ReadBoolean(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadBoolean
}
\keyword{internal}
