% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadCategoriseLegacy}
\alias{ReadCategoriseLegacy}
\title{ReadCategoriseLegacy}
\usage{
ReadCategoriseLegacy(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadCategoriseLegacy
}
\keyword{internal}
