% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadCategoriseList}
\alias{ReadCategoriseList}
\title{ReadCategoriseList}
\usage{
ReadCategoriseList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadCategoriseList
}
\keyword{internal}
