% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadCategoriseListLegacy}
\alias{ReadCategoriseListLegacy}
\title{ReadCategoriseListLegacy}
\usage{
ReadCategoriseListLegacy(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadCategoriseListLegacy
}
\keyword{internal}
