% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{readCharSafe}
\alias{readCharSafe}
\title{readCharSafe}
\usage{
readCharSafe(con, n, replace = " ")
}
\arguments{
\item{con}{A file connection - usually a binary file.}

\item{n}{The number of bytes to read.}

\item{replace}{a character to replace embedded nulls with.}
}
\value{
character vector.
}
\description{
reads `n` bytes as raw from a binary connection.
    Removes any embedded nuls, replacing them with `replace`.
}
\keyword{internal}
