% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadDataRecord}
\alias{ReadDataRecord}
\title{ReadDataRecord}
\usage{
ReadDataRecord(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadDataRecord
}
\keyword{internal}
