% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadDouble}
\alias{ReadDouble}
\title{ReadDouble}
\usage{
ReadDouble(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadDouble
}
\keyword{internal}
