% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadDoubleList}
\alias{ReadDoubleList}
\title{ReadDoubleList}
\usage{
ReadDoubleList(myFile)
}
\arguments{
\item{myFile}{An uncompressed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadDoubleList
}
\keyword{internal}
