% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadErrors}
\alias{ReadErrors}
\title{ReadErrors}
\usage{
ReadErrors(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadErrors
}
\keyword{internal}
