% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadFitList}
\alias{ReadFitList}
\title{ReadFitList}
\usage{
ReadFitList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadFitList
}
\keyword{internal}
