% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadGExportOptions}
\alias{ReadGExportOptions}
\title{ReadGExportOptions}
\usage{
ReadGExportOptions(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list of a single ExportOption (e.g., used in ExportXsi())
}
\description{
ReadGExportOptions
}
\keyword{internal}
