% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadGExportOptionsList}
\alias{ReadGExportOptionsList}
\title{ReadGExportOptionsList}
\usage{
ReadGExportOptionsList(myFile)
}
\arguments{
\item{myFile}{An uncompressed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list of ExportOptions
}
\description{
ReadGExportOptionsList
}
\keyword{internal}
