% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadGraph}
\alias{ReadGraph}
\title{ReadGraph}
\usage{
ReadGraph(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadGraph
}
\keyword{internal}
