% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadGroupsOneCase}
\alias{ReadGroupsOneCase}
\title{ReadGroupsOneCase}
\usage{
ReadGroupsOneCase(myFile, GroupVariables, AllVariables, CaseNum)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}

\item{GroupVariables}{A list of group variables for this case.}

\item{AllVariables}{A list of variables for this case.}

\item{CaseNum}{An integer representing the case number used in the lookup tables.}
}
\value{
A list
}
\description{
ReadGroupsOneCase
}
\keyword{internal}
