% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadHistory}
\alias{ReadHistory}
\title{ReadHistory}
\usage{
ReadHistory(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadHistory
}
\keyword{internal}
