% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadImplicitVar}
\alias{ReadImplicitVar}
\title{ReadImplicitVar}
\usage{
ReadImplicitVar(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadImplicitVar
}
\keyword{internal}
