% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadIntArray2}
\alias{ReadIntArray2}
\title{ReadIntArray2}
\usage{
ReadIntArray2(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A matrix
}
\description{
ReadIntArray2
}
\keyword{internal}
