% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadIntegerList}
\alias{ReadIntegerList}
\title{ReadIntegerList}
\usage{
ReadIntegerList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadIntegerList
}
\keyword{internal}
