% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadIntegerListList}
\alias{ReadIntegerListList}
\title{ReadIntegerListList}
\usage{
ReadIntegerListList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadIntegerListList
}
\keyword{internal}
