% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadIRecode}
\alias{ReadIRecode}
\title{ReadIRecode}
\usage{
ReadIRecode(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadIRecode
}
\keyword{internal}
