% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadIRecodeList}
\alias{ReadIRecodeList}
\title{ReadIRecodeList}
\usage{
ReadIRecodeList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadIRecodeList
}
\keyword{internal}
