% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadItemSet}
\alias{ReadItemSet}
\title{ReadItemSet}
\usage{
ReadItemSet(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadItemSet
}
\keyword{internal}
