% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadItemSetList}
\alias{ReadItemSetList}
\title{ReadItemSetList}
\usage{
ReadItemSetList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadItemSetList
}
\keyword{internal}
