% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadKeyList}
\alias{ReadKeyList}
\title{ReadKeyList}
\usage{
ReadKeyList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadKeyList
}
\keyword{internal}
