% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadLookUpList}
\alias{ReadLookUpList}
\title{ReadLookUpList}
\usage{
ReadLookUpList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadLookUpList
}
\keyword{internal}
