% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadMatrix}
\alias{ReadMatrix}
\title{ReadMatrix}
\usage{
ReadMatrix(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadMatrix
}
\keyword{internal}
