% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadMatrix_C}
\alias{ReadMatrix_C}
\title{ReadMatrix_C}
\usage{
ReadMatrix_C(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadMatrix_C
}
\keyword{internal}
