% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadMatrixList}
\alias{ReadMatrixList}
\title{ReadMatrixList}
\usage{
ReadMatrixList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadMatrixList
}
\keyword{internal}
