% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadNamedStringList}
\alias{ReadNamedStringList}
\title{ReadNamedStringList}
\usage{
ReadNamedStringList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A data frame
}
\description{
ReadNamedStringList
}
\keyword{internal}
