% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadRandomStructure}
\alias{ReadRandomStructure}
\title{ReadRandomStructure}
\usage{
ReadRandomStructure(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadRandomStructure
}
\keyword{internal}
