% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadRegression}
\alias{ReadRegression}
\title{ReadRegression}
\usage{
ReadRegression(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadRegression
}
\keyword{internal}
