% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadRegressionList}
\alias{ReadRegressionList}
\title{ReadRegressionList}
\usage{
ReadRegressionList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadRegressionList
}
\keyword{internal}
