% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadRegressionListLeg}
\alias{ReadRegressionListLeg}
\title{ReadRegressionListLeg}
\usage{
ReadRegressionListLeg(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadRegressionListLeg
}
\keyword{internal}
