% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadRegressors}
\alias{ReadRegressors}
\title{ReadRegressors}
\usage{
ReadRegressors(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadRegressors
}
\keyword{internal}
