% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadResponseList}
\alias{ReadResponseList}
\title{ReadResponseList}
\usage{
ReadResponseList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadResponseList
}
\keyword{internal}
