% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadStringList}
\alias{ReadStringList}
\title{ReadStringList}
\usage{
ReadStringList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadStringList
}
\keyword{internal}
