% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState.R
\name{ReadSys}
\alias{ReadSys}
\title{ReadSys}
\usage{
ReadSys(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file created by the `put` command in 'ACER ConQuest'. The put command must use the option `compressed = no`.}
}
\value{
A list containing the data objects created by 'ACER ConQuest'.
}
\description{
Internal function to read an 'ACER ConQuest' system file. Called by conquestr::ConQuestSys.
}
\seealso{
conquestr::ConQuestSys()
}
