% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadTermsList}
\alias{ReadTermsList}
\title{ReadTermsList}
\usage{
ReadTermsList(myFile, cqs_version)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadTermsList
}
\keyword{internal}
