% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadVariable}
\alias{ReadVariable}
\title{ReadVariable}
\usage{
ReadVariable(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadVariable
}
\keyword{internal}
