% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadVariableList}
\alias{ReadVariableList}
\title{ReadVariableList}
\usage{
ReadVariableList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadVariableList
}
\keyword{internal}
