% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadVarList}
\alias{ReadVarList}
\title{ReadVarList}
\usage{
ReadVarList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadVarList
}
\keyword{internal}
