% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadYOneCase}
\alias{ReadYOneCase}
\title{ReadYOneCase}
\usage{
ReadYOneCase(myFile, NReg)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}

\item{NReg}{An integer representing the number of regressors in the model, a representation of 'ACER ConQuest' object gNReg.}
}
\value{
A list
}
\description{
ReadYOneCase
}
\keyword{internal}
