% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningHelpers.R
\name{recodeResps}
\alias{recodeResps}
\title{recodeResps}
\usage{
recodeResps(data, recodeMap, varLabel, rawLabel, recodeLabel)
}
\arguments{
\item{data}{Raw data, a data frame.}

\item{recodeMap}{A data frame which contains the raw responses and corresponding recoded responses
for each of the items in long form.}

\item{varLabel}{A variable name in recodeMap that identifes the item label.}

\item{rawLabel}{A variable name in recodeMap that identifies the raw item responses to be recoded.}

\item{recodeLabel}{A variable name in recodeMap that idenitfies the new values to recode to.}
}
\value{
a data frame with raw data recoded according to recodeMap.
}
\description{
Recode raw item responses for analyses.
}
