% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replaceInDataFrame}
\alias{replaceInDataFrame}
\title{iterate through a data frame and use replaceInVector}
\usage{
replaceInDataFrame(d, r, x)
}
\arguments{
\item{d}{A DataFrame.}

\item{r}{A double - the value to be replaced if it is < -1e300.}

\item{x}{A double - the value to repalce r with.}
}
\description{
iterate through a data frame and use replaceInVector
}
