% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replaceInVector}
\alias{replaceInVector}
\title{replace a very large neagtive number with something - usually NA_REAL}
\usage{
replaceInVector(v, r, x)
}
\arguments{
\item{v}{A NumericVector.}

\item{r}{A double - the value to be replaced if it is < -1e300.}

\item{x}{A double - the value to repalce r with.}
}
\description{
replace a very large neagtive number with something - usually NA_REAL
}
