% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestRout_createDF.R
\name{routPointsToDf}
\alias{routPointsToDf}
\title{routPointsToDf}
\usage{
routPointsToDf(myRout)
}
\arguments{
\item{myRout}{An 'ACER ConQuest' Rout file created by a call to 'plot' in 'ACER ConQuest'.}
}
\value{
A data frame containing the series that make up the plot
}
\description{
routPointsToDf
}
\keyword{internal}
