% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestRout_createDF.R
\name{routType}
\alias{routType}
\title{routType}
\usage{
routType(myRout)
}
\arguments{
\item{myRout}{An 'ACER ConQuest' Rout file created by a call to 'plot' in 'ACER ConQuest'.}
}
\value{
A string naming the type of plot
}
\description{
routType
}
\keyword{internal}
