% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{searchConQuestSys}
\alias{searchConQuestSys}
\title{searchConQuestSys}
\usage{
searchConQuestSys(searchString, mySys, value = TRUE, ignore.case = TRUE)
}
\arguments{
\item{searchString}{A string to search within the names of mySys.}

\item{mySys}{An 'ACER ConQuest' system file object created using the conquestr::ConQuestSys function.}

\item{value}{Should searchConQuestSys return the name of the object or its index.}

\item{ignore.case}{Should searchConQuestSys ignore the case of the search term.}
}
\value{
a string including object names mathching the search term
}
\description{
Search for object names within a ConQuest System file object.
}
