% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{short_test}
\alias{short_test}
\title{Unscored student response data.}
\format{A tibble with 1475 rows and 10 variables:
\describe{
  \item{pid}{a unique person identifier}
  \item{q1:q9}{q1 to q9 are multiple choice test items, \code{NA} indicates no response was provided}
}}
\source{
\url{http://www.smeebu.com/}
}
\usage{
short_test
}
\description{
A dataset containing the responses of 1000 students to 9 multiple choice test items
}
\keyword{datasets}
