% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{short_test_score_key}
\alias{short_test_score_key}
\title{Correct answer key for \code{data(short_test)}}
\format{A tibble with 9 rows and 2 variables:
\describe{
  \item{item}{a unique item id}
  \item{resp}{the correct resp value for the corresponding item}
}}
\source{
\url{http://www.smeebu.com/}
}
\usage{
short_test_score_key
}
\description{
A tibble containing the correst answer option for the responses provided in \code{data(short_test)}
}
\keyword{datasets}
