% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{summariseCqChain}
\alias{summariseCqChain}
\title{summariseCqChain}
\usage{
summariseCqChain(myChain)
}
\arguments{
\item{myChain}{A data frame returned from getCqChain.}
}
\value{
A list.
}
\description{
takes a data frame created by getCqChain and returns a list reporting the mean and variaince for each parameter
}
\examples{
\dontrun{
summariseCqChain(getCqChain(ConQuestSys()))
}
}
