% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showHelpers.R
\name{sysFileOk}
\alias{sysFileOk}
\title{sysFileOk}
\usage{
sysFileOk(sysFile, defaultSys)
}
\arguments{
\item{sysFile}{An ACER ConQuest system file read into R using conquestr::ConQuestSys}

\item{defaultSys}{A Boolean indicating if sysFile is the defualt system file
created by an empty call to conquestr::ConQuestSys}
}
\description{
checks
}
\examples{
\dontrun{
sysFileOkResult <- sysFileOk(conquestr::ConQuestSys())
}
}
