% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{sysToDimLabels}
\alias{sysToDimLabels}
\title{sysToDimLabels}
\usage{
sysToDimLabels(mySys, myWarn = TRUE)
}
\arguments{
\item{mySys}{An R object of class ConQuestSys, returned by the function conquestr::ConQuestSys.}

\item{myWarn}{a bool indicating whether a warning should be printed if there are no dimension labels.}
}
\value{
a data frame.
}
\description{
Gets dimensions labels from a ConQuest system file.
}
\seealso{
conquestr::sysToBMatrixDf()
}
\keyword{internal}
