% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{sysToItemLabels}
\alias{sysToItemLabels}
\title{sysToItemLabels}
\usage{
sysToItemLabels(mySys, myWarn = TRUE)
}
\arguments{
\item{mySys}{An R object of class ConQuestSys, returned by the function conquestr::ConQuestSys.}

\item{myWarn}{a bool indicating whether a warning should be printed if there are no item labels.}
}
\value{
a data frame.
}
\description{
Gets item labels from a ConQuest system file.
}
\seealso{
conquestr::sysToBMatrixDf()
}
\keyword{internal}
