% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneral.R
\name{sysToItems}
\alias{sysToItems}
\title{sysToItems}
\usage{
sysToItems(mySys, myDims)
}
\arguments{
\item{mySys}{An R object of class ConQuestSys, returned by the function conquestr::ConQuestSys}

\item{myDims}{A string specifying if all or specific dimensions should be included. The default is "all", Specific dimensions are specified by the label "D1" for dimensions 1 etc.}
}
\value{
A data frame containing R the labelled B matrix.
}
\description{
Read an R object of class ConQuestSys and create a labelled representation of the B matrix (scoring matrix). This maps item response catagories to items and dimensions.
}
\seealso{
conquestr::sysToItemDifDf()
}
\keyword{internal}
