% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHelpers.R
\name{thetaScore}
\alias{thetaScore}
\title{thetaScore}
\usage{
thetaScore(theta, responses, itemParams, perfAdjust = 0.3)
}
\arguments{
\item{theta}{a scalar value of theta.}

\item{responses}{a vector of item responses for one case (can contain NA). Responses are in the same order as `items`}

\item{itemParams}{a list of item designs (each used in a call to `simplef`). Must be of same length as `responses`.}

\item{perfAdjust}{the correction factor for zeros and perfects to be add/subtracted from the raw score.}
}
\value{
a double, the score (O-E) at theta.
}
\description{
Return score (O-E) at theta. Used in estimation of theta by minimising _raw score - expected score_ at theta
}
\keyword{internal}
