% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{transformPvs}
\alias{transformPvs}
\title{transformPvs}
\usage{
transformPvs(data, mT = 0, sdT = 1, weights = 1)
}
\arguments{
\item{data}{A data frame or matrix that contains the PVs}

\item{mT}{The desired mean of the PVs}

\item{sdT}{The desired sd of the PVs}

\item{weights}{a vector of weights, the same length as \code{data[1]} used to
calculate the mean and SD across the PVs}
}
\value{
a List of transformed PVs with as many elements as PVs were listed in 'x'.
}
\description{
Helper function to Transform PVs onto a new metric
(e.g., PISA Mean = 500, SD = 100).
Uses the method described in the PISA 2012 technical manual.
}
