% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{zapNulls}
\alias{zapNulls}
\title{zapNulls}
\usage{
zapNulls(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
x.
}
\description{
Zaps NULL values embedded in ConQuest data objects.
}
\keyword{internal}
