% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{zapSystemMissing}
\alias{zapSystemMissing}
\title{zapSystemMissing}
\usage{
zapSystemMissing(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
x
}
\description{
Coerce ConQuest system missing values to NA. Note this is very slow and users should use the internal function conquestr::replaceInDataFrame where possible.
}
\keyword{internal}
