% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestr.R
\name{ConQuestCall}
\alias{ConQuestCall}
\title{ConQuestCall}
\usage{
ConQuestCall(cqInstallLocation, cqc, stdout = "")
}
\arguments{
\item{cqInstallLocation}{The location of the 'ACER ConQuest' executable.}

\item{cqc}{The locaiton of the control file to be run.}

\item{stdout}{On Mac only, can be toggled to NULL (or a connection) to supress output to R console.}
}
\value{
prints 'ACER ConQuest' output to stdout.
}
\description{
Call 'ACER ConQuest' and run a control file.
}
\examples{
\dontrun{
ConQuestCall(cqInstallLocation = file.path("/Applications", "ConQuest BETA", "ConQuest"))
}
}
