% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestr.R
\name{ConQuestRout}
\alias{ConQuestRout}
\title{ConQuestRout}
\usage{
ConQuestRout(myRout)
}
\arguments{
\item{myRout}{The location of an 'ACER ConQuest' rout file created by 'ACER ConQuest' > 5.1.4.}
}
\value{
A list containing the data objects created by 'ACER ConQuest' plot command.
}
\description{
Read an ''ACER ConQuest'' rout file created by a `plot` command in 'ACER ConQuest'.
}
\examples{
myPlot<- ConQuestRout()
\dontrun{
# if you run the above example you will have the points from a plot ICC command.
str(myPlot)
}
}
