% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_itanal.R
\name{cq_example_itanal}
\alias{cq_example_itanal}
\title{Example ConQuest item analysis file}
\usage{
cq_example_itanal(n = 10, display = TRUE)
}
\arguments{
\item{n}{the number of lines to print. only used when \code{display = TRUE}.}

\item{display}{When \code{TRUE} (the default) the example itanal is read and the first \code{n} lines are printed.
The file path to example itanal file is returned invisibly.
Use \code{FALSE} if just the itanal file path is needed. No output will be displayed.}
}
\value{
Invisibly returns a valid file path to an itanal file which can be used with other conquestr functions.
}
\description{
Example ConQuest item analysis file
}
\examples{
library(readr)
fname <- cq_example_itanal(display = FALSE)
fname
cq_example_itanal(n = 7)

}
