% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_show.R
\name{cq_show}
\alias{cq_show}
\title{Read in ConQuest show output file}
\usage{
cq_show(fname, fit = c(TRUE, FALSE))
}
\arguments{
\item{fname}{A file path to an existing ConQuest show file (e.g.
'myfile.shw'). Can take other input as for \code{readr::read_file}}

\item{fit}{A logical value indicating if the show file has fit statistics
included or not. This only affects warnings generated. It is convenient
when no fit statistics have been requested as the default value \code{TRUE}
generates warnings that data is missing.}
}
\description{
Read in ConQuest show output file
}
\examples{
fname <- cq_example(display = FALSE, example_name = "ex1.shw")
df <- cq_show(fname)

}
