% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState_createDF.R
\name{createDfFromSys}
\alias{createDfFromSys}
\title{createDfFromSys}
\usage{
createDfFromSys(mySys)
}
\arguments{
\item{mySys}{An R object of class ConQuestSys, returned by the function conquestr::ConQuestSys}
}
\value{
A list containing R data frames based on the list objects in the ConQuest system file that has been read in.
}
\description{
Internal function to read an R object of class ConQuestSys (which returns a list) and create neat R data frame objects. Called by conquestr::ConQuestSysDf.
}
\seealso{
conquestr::ConQuestSys()
}
