% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqChain}
\alias{getCqChain}
\title{getCqChain}
\usage{
getCqChain(myCqs)
}
\arguments{
\item{myCqs}{A system file.}
}
\value{
A data frame.
}
\description{
creates a data frame representation of the estimation chain from an MCMC model.
The burn is discarded and only the unskipped itterations in MCMC chain are retained.
}
\examples{
\dontrun{
getCqChain(ConQuestSys())
}
}
