% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState_createDF.R
\name{getCqData}
\alias{getCqData}
\title{getCqData}
\usage{
getCqData(mySys)
}
\arguments{
\item{mySys}{An R object of class ConQuestSys, returned by the function conquestr::ConQuestSys}
}
\value{
A List of data frames.
}
\description{
Get data objects from an R object of class ConQuestSys.
    This function returns person IDs, response data, case estimates, regression and weight data.
    Each data type is stored as a data frame, and each data frame is a named element of a list.
    1. PID, 2. Responses, 3. Estimates, 4. Regression.
}
\examples{
mySys <- ConQuestSys()
myData <- getCqData(mySys)
}
\seealso{
conquestr::ConQuestSys()
}
