% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquestrFunc.R
\name{getCqParams}
\alias{getCqParams}
\title{getCqParams}
\usage{
getCqParams(sysFile)
}
\arguments{
\item{sysFile}{An ACER ConQuest system file read into R using conquestr::ConQuestSys.}
}
\value{
A data frame.
}
\description{
creates a data frame representation of the parameters of the model,
including both estimated, constrained, and anchored parameters.
}
\examples{
\dontrun{
getCqParams(ConQuestSys())
}
}
\keyword{internal}
