% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHelpers.R
\name{itemResp}
\alias{itemResp}
\title{itemResp}
\usage{
itemResp(myAblty, myItem)
}
\arguments{
\item{myAblty}{A numeric ability.}

\item{myItem}{A single item (a matrix k categories by three (categor score, delta dot, tau).}
}
\value{
A numeric score, usually an integer in the range 0, 1, 2, ..., k-1.
}
\description{
Returns a probabilistic item response, given one ability and one item (a matrix k categories by three (categor score, delta dot, tau).
}
\examples{
myItem<- matrix(c(0, 0, 0, 1, 1, 0), ncol = 3, byrow = TRUE)
itemResp(0, myItem)
}
\keyword{internal}
