% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoHelpers.R
\name{myItemHelper}
\alias{myItemHelper}
\title{myAbilitiesHelper}
\usage{
myItemHelper(myWhat = "ok", myItems = NA)
}
\arguments{
\item{myWhat}{A character string indicating what to check.}

\item{myItems}{A numeric vector of item deltas.}
}
\value{
A Boolean or a double.
}
\description{
helper function to return attributes of vector of item deltas for use in information function and WI.
}
\keyword{internal}
