% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHelpers.R
\name{perturbProbs}
\alias{perturbProbs}
\title{perturbProbs}
\usage{
perturbProbs(myProbs, perturbP)
}
\arguments{
\item{myProbs}{A matrix of response probabilities from n persons * i items.}

\item{perturbP}{A matrix of response probabilities from n persons * i items.}
}
\value{
A matrix.
}
\description{
Perturbs response probabilities.
}
\examples{
myExProbs<- matrix(rnorm(100, 0, 1), 10, 10)
myExPerturb<- list(data.frame(item = 1, type = "flat", pivot = 0.5, factor = 25))
myProbs<- perturbProbs(myProbs = myExProbs, perturbP = myExPerturb)
}
\keyword{internal}
