% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRout.R
\name{plotRout}
\alias{plotRout}
\alias{plotRout.TestInfo}
\alias{plotRout.InformationWithLatentDist}
\alias{plotRout.ICC}
\alias{plotRout.MCC}
\alias{plotRout.TCC}
\alias{plotRout.default}
\title{plotRout}
\usage{
plotRout(myRout, ...)

\method{plotRout}{TestInfo}(myRout, ...)

\method{plotRout}{InformationWithLatentDist}(myRout, ...)

\method{plotRout}{ICC}(myRout, ...)

\method{plotRout}{MCC}(myRout, ...)

\method{plotRout}{TCC}(myRout, ...)

\method{plotRout}{default}(myRout, ...)
}
\arguments{
\item{myRout}{an R object created by the \code{ConQuestRout} function.}

\item{...}{additional arguments passed into plotting functions}
}
\value{
A ggplot2 object.
}
\description{
generates a plot from an 'ACER ConQuest' Rout file. use \code{ConQuestRout} to read in an Rout
file created by a \code{plot} command in 'ACER ConQuest'.
}
\examples{
myRout <- ConQuestRout()
myPlot <- plotRout(myRout)
## to see why we import this, see https://ggplot2.tidyverse.org/articles/ggplot2-in-packages.html
}
