% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadAllResponseData}
\alias{ReadAllResponseData}
\title{ReadIntegerList}
\usage{
ReadAllResponseData(myFile, N)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}

\item{N}{An integer representation of 'ACER ConQuest' object gNCases.}
}
\value{
A list
}
\description{
ReadIntegerList
}
\keyword{internal}
