% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadAnchorList}
\alias{ReadAnchorList}
\title{ReadAnchorList}
\usage{
ReadAnchorList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadAnchorList
}
\keyword{internal}
