% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadBandDefine}
\alias{ReadBandDefine}
\title{ReadBandDefine}
\usage{
ReadBandDefine(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadBandDefine
}
\keyword{internal}
