% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadBitSet}
\alias{ReadBitSet}
\title{ReadBitSet}
\usage{
ReadBitSet(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadBitSet
}
\keyword{internal}
