% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadCategorise}
\alias{ReadCategorise}
\title{ReadCategorise}
\usage{
ReadCategorise(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadCategorise
}
\keyword{internal}
