% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadCodeList}
\alias{ReadCodeList}
\title{ReadCodeList}
\usage{
ReadCodeList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadCodeList
}
\keyword{internal}
