% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadEstimatesRecord}
\alias{ReadEstimatesRecord}
\title{ReadEstimatesRecord}
\usage{
ReadEstimatesRecord(myFile, Dimensions, NPlausibles, n)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}

\item{Dimensions}{An integer representation of 'ACER ConQuest' object gNDim.}

\item{NPlausibles}{An integer representation of 'ACER ConQuest' object gNPlausibles.}

\item{n}{An integer representation of 'ACER ConQuest' object gNCases.}
}
\value{
A list
}
\description{
ReadEstimatesRecord
}
\keyword{internal}
