% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadInteger}
\alias{ReadInteger}
\title{ReadInteger}
\usage{
ReadInteger(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadInteger
}
\keyword{internal}
