% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadIntegerList}
\alias{ReadIntegerList}
\title{ReadIntegerList}
\usage{
ReadIntegerList(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadIntegerList
}
\keyword{internal}
