% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadLabel}
\alias{ReadLabel}
\title{ReadLabel}
\usage{
ReadLabel(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadLabel
}
\keyword{internal}
