% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadMatrixVars}
\alias{ReadMatrixVars}
\title{ReadMatrixVars}
\usage{
ReadMatrixVars(myFile)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' system file.}
}
\value{
A list
}
\description{
ReadMatrixVars
}
\keyword{internal}
