% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadParametersList}
\alias{ReadParametersList}
\title{ReadParametersList}
\usage{
ReadParametersList(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadParametersList
}
\keyword{internal}
