% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadResponse}
\alias{ReadResponse}
\title{ReadResponse}
\usage{
ReadResponse(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadResponse
}
\keyword{internal}
