% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadSeries}
\alias{ReadSeries}
\title{ReadSeries}
\usage{
ReadSeries(myFile)
}
\arguments{
\item{myFile}{An uncompresed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadSeries
}
\keyword{internal}
