% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState.R
\name{ReadSys}
\alias{ReadSys}
\title{ReadSys}
\usage{
ReadSys(myFile, isMini)
}
\arguments{
\item{myFile}{An connection to an 'ACER ConQuest' system file created by the
\code{put} command in 'ACER ConQuest'.
If the file is compressed, and uncompressed temporary file is created.}

\item{isMini}{A boolean, set to TRUE if the system file is a \emph{mini} system file created by 'ACER ConQuest'
command put with option "mini = yes".}
}
\value{
A list containing the data objects created by 'ACER ConQuest'.
}
\description{
Internal function to read an 'ACER ConQuest' system file.
Called by conquestr::ConQuestSys.
}
\seealso{
conquestr::ConQuestSys()
}
