% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestState.R
\name{ReadSysMini}
\alias{ReadSysMini}
\title{ReadSysMini}
\usage{
ReadSysMini(myFile, Dimensions, N, NPlausibles, isDebug)
}
\arguments{
\item{myFile}{An 'ACER ConQuest' \emph{mini} system file created by the \code{put}
command in 'ACER ConQuest' with the option "mini = yes".}

\item{Dimensions}{gNDim object passed in to this call to \code{ReadSysMini} from
the higher-level original call to ReadSys. This value is returned in the
list returned by this function call.}

\item{N}{gNCases object passed in to this call to \code{ReadSysMini} from
the higher-level original call to ReadSys. This value is returned in the
list returned by this function call.}

\item{NPlausibles}{gNPlausibles object passed in to this call to \code{ReadSysMini} from
the higher-level original call to ReadSys. This value is returned in the
list returned by this function call.}

\item{isDebug}{Bool. Passed in to this call to \code{ReadSysMini} from
the higher-level original call to ReadSys. This value is used to trigger
output of debug information to standard out.}
}
\value{
A list containing the data objects created by 'ACER ConQuest'.
}
\description{
Internal function to read an 'ACER ConQuest' system file.
Called by conquestr::ConQuestSys.
}
\seealso{
conquestr::ConQuestSys()
}
