% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadConQuestLibrary.R
\name{ReadTerms}
\alias{ReadTerms}
\title{ReadTerms}
\usage{
ReadTerms(myFile, cqs_version)
}
\arguments{
\item{myFile}{An uncompressed 'ACER ConQuest' system file created by 'ACER ConQuest'.}
}
\value{
A list
}
\description{
ReadTerms
}
\keyword{internal}
